add problens in make a list a personalTrainer
