+++ client/src/sample/controllers/RegisterTrainningPlanController.java
@@ -16,6 +16,8 @@ import java.io.IOException;
 
 import javafx.scene.control.Alert.AlertType;
 
+import static sample.utils.Utils.onlyDigits;
+
 public class RegisterTrainningPlanController {
 
 
@@ -50,17 +52,19 @@ public class RegisterTrainningPlanController {
             Alert alert = new Alert(AlertType.ERROR);
             alert.setHeaderText("O preço do Plano de treino nao foi inserido!");
             alert.showAndWait();
-        }
-        // System.out.println("nome: " +  name.getText() + " descrition: " + descrition.getText() + " float: " +amount.getText());
-
-        if (!name.getText().isEmpty() && !descrition.getText().isEmpty() && !amount.getText().isEmpty()) {
-    //        System.out.println("nome: " + name.getText() + " descrition: " + descrition.getText() + " float: " + amount.getText());
-            final   PlanoTreino planoTreino = new PlanoTreino(name.getText(), descrition.getText(), Integer.parseInt(amount.getText()));
-            Datasource.getInstance().registerPlanoTreino(planoTreino.getName(), planoTreino.getDescrition(), planoTreino.getAmount());
-            Alert alert = new Alert(AlertType.INFORMATION);
-            alert.setHeaderText("O registro foi feito com sucesso");
-            alert.showAndWait();
-
+        } else {
+            if(onlyDigits(amount.getText())) {
+                //System.out.println("nome: " + name.getText() + " descrition: " + descrition.getText() + " float: " + amount.getText());
+                final PlanoTreino planoTreino = new PlanoTreino(name.getText(), descrition.getText(), Integer.parseInt(amount.getText()));
+                Datasource.getInstance().registerPlanoTreino(planoTreino.getName(), planoTreino.getDescrition(), planoTreino.getAmount());
+                Alert alert = new Alert(AlertType.INFORMATION);
+                alert.setHeaderText("O registro foi feito com sucesso");
+                alert.showAndWait();
+            }else{
+                Alert alert = new Alert(Alert.AlertType.ERROR);
+                alert.setHeaderText("Insira-o montante incorretamente");
+                alert.showAndWait();
+            }
 
 
             Main.setRoot("listTrainningPlan");
