--- client/src/sample/controllers/Utils.java
@@ -1,4 +1,0 @@
-package sample.controllers;
-
-public class Utils {
-}
